adding follwed tags api, moving some extension function to their proper place
