+++ core/repository/mastodon/src/main/java/social/firefly/core/repository/mastodon/PollRepository.kt
@@ -7,6 +7,7 @@ import social.firefly.core.database.model.entities.DatabasePoll
 import social.firefly.core.model.Emoji
 import social.firefly.core.model.Poll
 import social.firefly.core.model.PollOption
+import social.firefly.core.repository.mastodon.model.poll.toDatabaseModel
 
 class PollRepository internal constructor(private val dao: PollsDao) {
 
@@ -30,32 +31,3 @@ class PollRepository internal constructor(private val dao: PollsDao) {
 
     suspend fun deleteOldPolls() = dao.deleteOldPolls()
 }
-
-private fun Poll.toDatabaseModel(): DatabasePoll =
-    DatabasePoll(
-        pollId = pollId,
-        isExpired = isExpired,
-        allowsMultipleChoices = allowsMultipleChoices,
-        votesCount = votesCount,
-        options = options.map { it.toDatabaseModel() },
-        emojis = emojis.map { it.toDatabaseModel() },
-        expiresAt = expiresAt,
-        votersCount = votersCount,
-        hasVoted = hasVoted,
-        ownVotes = ownVotes,
-    )
-
-fun Emoji.toDatabaseModel(): DatabaseEmoji =
-    DatabaseEmoji(
-        shortCode = shortCode,
-        url = url,
-        staticUrl = staticUrl,
-        isVisibleInPicker = isVisibleInPicker,
-        category = category,
-    )
-
-fun PollOption.toDatabaseModel(): DatabasePollOption =
-    DatabasePollOption(
-        title = title,
-        votesCount = votesCount,
-    )
