+++ core/repository/mastodon/src/main/java/social/firefly/core/repository/mastodon/model/status/ExternalToDatabase.kt
@@ -12,6 +12,7 @@ import social.firefly.core.database.model.DatabaseMention
 import social.firefly.core.database.model.DatabaseSource
 import social.firefly.core.database.model.DatabaseStatusVisibility
 import social.firefly.core.database.model.entities.DatabaseAccount
+import social.firefly.core.database.model.entities.DatabaseStatus
 import social.firefly.core.database.model.entities.accountCollections.DatabaseBlock
 import social.firefly.core.database.model.entities.accountCollections.DatabaseMute
 import social.firefly.core.model.Account
@@ -25,7 +26,44 @@ import social.firefly.core.model.FocalPoint
 import social.firefly.core.model.History
 import social.firefly.core.model.Mention
 import social.firefly.core.model.Source
+import social.firefly.core.model.Status
 import social.firefly.core.model.StatusVisibility
+
+fun Status.toDatabaseModel(): DatabaseStatus =
+    DatabaseStatus(
+        statusId = statusId,
+        uri = uri,
+        createdAt = createdAt,
+        accountId = account.accountId,
+        content = content,
+        visibility = visibility.toDatabaseModel(),
+        isSensitive = isSensitive,
+        contentWarningText = contentWarningText,
+        mediaAttachments = mediaAttachments.map { it.toDatabaseModel() },
+        mentions = mentions.map { it.toDatabaseModel() },
+        hashTags = hashTags.map { it.toDatabaseModel() },
+        emojis = emojis.map { it.toDatabaseModel() },
+        boostsCount = boostsCount,
+        favouritesCount = favouritesCount,
+        repliesCount = repliesCount,
+        application = application?.toDatabaseModel(),
+        url = url,
+        inReplyToId = inReplyToId,
+        inReplyToAccountId = inReplyToAccountId,
+        inReplyToAccountName = inReplyToAccountName,
+        boostedStatusId = boostedStatus?.statusId,
+        boostedStatusAccountId = boostedStatus?.account?.accountId,
+        pollId = poll?.pollId,
+        card = card?.toDatabaseModel(),
+        language = language,
+        plainText = plainText,
+        isFavorited = isFavourited,
+        isBoosted = isBoosted,
+        isMuted = isMuted,
+        isBookmarked = isBookmarked,
+        isPinned = isPinned,
+        isBeingDeleted = isBeingDeleted,
+    )
 
 fun Account.toDatabaseModel(): DatabaseAccount =
     DatabaseAccount(
